don't rely on /proc for 'browse' mode
Instead, pass the script code via stdin.
We end up with a zombie (the child process that passes the script code)
because we can't ignore SIGCHLD in the parent (the Python code relies on
being able to wait for children), but the zombie dies with the outer
process.  This probably could be avoided by double-forking or other
approaches, but it doesn't seem worth the effort.