If a command fails, wait for all running commands to terminate before we do
Previously, if a command fails, the fate of the other child processes
running in parallel was inadequately controlled.  On POSIX platforms,
the processes were orphaned.  Normally they would run to completion,
but were liable to being killed by a SIGPIPE.  On Windows, the child
processes would terminate with the parent.  The cleanup-on-interrupt
patch caused the SubprocessSet and Builder destructors to clean
up after themselves by killing any running child processes and
deleting their output files, making the behaviour more predictable
and consistent across platforms.

If the build is interrupted by the user, this is correct behaviour.
But in the case where the build is stopped by a failed command, this
would be inconsistent with user expectations.  In the latter case,
we now let any remaining child processes run to completion before
leaving the main loop in Builder::Build.