Treat paths of the form "existing-file/something" as non-existent
Some people like to construct phony target names by appending a
"/something" suffix to an existing target "foo".  But if "foo" is an
existing file, stat will report ENOTDIR for this path, causing ninja
to spew errors.  Fix this by treating ENOTDIR in the same way as we
do ENOENT -- as a non-existent path.