Add -l N option to limit the load average.
This is similar to GNU make -l/--load-average option.  It limits the number
of job started if the load average exceed the given value.  It can be very
useful when running ninja on a continuous integration server where we want
to use parallelism as much as possible without overloading the server.