The big realtime patch (use 100ns ticks)
Dependencies are tracked with 100ns resolution.

The timestamps from filesystem and the current ticks are now realtime.
That means, that if a dependency file change at same sec as the last
target build it will be found and the needed rebuild is triggered.

Too the mtime stamp at build log is updateted by the way and will be
commpared if needed to support restat with higher resolution.