disallow crlf in manifest files
It turns out to be trickier than expected to process these correctly.
It turns out to also be trickier than expected to give a nice error
message on encountering these.  But the behavior prior to this patch
would just be silent failures where we attempted to examine paths
that accidentally contained embedded \r.

For now, fix all regexes of the form [^...] to include \r in the
excluded block, then assert that we get a vague lexer error near the
problem.

In the future perhaps we can open manifest files in text mode on Windows
or just disallow Windows-style CRLF in the manual.