Implement a command output flush mechanism.
Ninja's short output status is fine 99% of the time.  However for long command
generating a long output like documentation generation, configuration stage
or tests execution it is not really convenient.  This patch fills this gap
by introducing the rule attribute "flush" which, if present, causes the output
of the command to be flushed if not executed in parallel.