Implement a command output flush mechanism.
Ninja's short output status is fine 99% of the time. However for command
lasting a long time and generating a long output like documentation
generation, configuration stage or tests execution it is not really
convenient. This patch fills this gap by automatically flushing the output
when we are sure that no other command can interleave their output.