don't use PATH_MAX
There is not a single system where this #define actually works:

* modern unices don't have real limits here, and set PATH_MAX to an
  arbitrary value (4096 or 1024)
* Windows defines _MAX_PATH to 260, yet allows paths up to around 32KB
  Unicode characters if you prefix them with \\?\.  For added horror,
  the latter limit is "approximate", according to
  http://msdn.microsoft.com/en-us/library/aa365247.aspx
* Hurd guys decided to weed out truncation bugs by force, and adamantly
  refuse to include this #define for compat with broken code

It can be argued that scenarios for hitting this limit in real life are
quite convoluted, yet there is no reason to not do this right.
My real reason, though, is that Debian sends annoying mails that a
build failed on Debian/Hurd, so let's get this cleared.