configure.py/bootstrap.py — account for CXXFLAGS.
Take into account CXXFLAGS in the environment. I've tried to ensure
this doesn't interfere with existing workflows which depend on using
CFLAGS only by appending rather than replacing CFLAGS; this may lead to
some extraneous duplicates, but it seems better than the alternative.
Also, CXXFLAGS (and CFLAGS) are added to LDFLAGS with this change; this
is to account for some special flags (like clang's -stdlib) which
modify the behavior of both the compiler and the linker.