Improve ninja responsiveness to SIGINT.
This problem is subtle. When executing many short (in time) sub-processes,
ninja does not respond well to SIGINT. This due to the short time frame where
ninja is not blocking SIGINT. Before this patch the time frame was the time
frame of the pselect(1)/ppoll(1) system call. This time frame get very short
when sub-processes does not live for long (the worst case is with "sleep 0"
command). With this patch it is no longer an issue.