Allow to stop ninja and its sub-processes at once.
Before this patch each sub-processes had its own group. This was great
because we could terminate each sub-processes group independently. Thus, author
of commands started by ninja did not have to worry about terminating their
own sub-processes. They were part of the same group as their command and ninja
was terminating the whole group.

However, with this approach the ninja process and its direct sub-processes
are not in the same group. This is a problem when you want to stop (= sending
SIGSTOP or hitting control-z) or continue a ninja job. In this case only the
ninja process is stopped and its direct sub-processes keep running until they
block on their IO because ninja is no longer reading them. If they do not
write, then they are not stopped at all.

With this patch you can use control-z but signal is sent to each sub-process
instead of each sub-process's group.

make(1) has the same behavior.