Solve the line-wrapping performance problem by shortening lines.
Avoid joining a list just to split it again.

The main performance problem in the line wrapping came from lines that were
thousands of characters long, and those came from a single location, build.

This patch jumps past _line and the slow code, by emitting the file
list with one file per line. It even becomes easier to read that way.
Everyone is a winner.

(Not wrapping at all is slightly faster but in the chromium project,
 only by 0.1 second or so).