Process the -C option before running RUN_AFTER_FLAGS tools.
This brings the flag more inline with its description: "change to
DIR before doing anything else".

The use case is to make it possible to use -C together with -t msvc.
When debugging Windows builds, it's handy to be able to copy-paste
the commands from "ninja -v" and make them run in the correct
directory by adding this flag.