Allow duplicate rules with the same content
Ninja till now disallowed duplicate rules with the same name
in the global namespace. With this change, ninja allows duplicate
rules with *exactly* the same content. The content check is done
by computing and comparing the hash values of the rules' content.

This change was tested using the following compilers:
 - MSVC 18.00.21005.1 for x64
 - g++ 4.9.2