Use waitid to check done-ness of console processes
For console pool processes, rather than checking the
closure of the pipe, use "waitid" to check if the
process has exited. We specify WNOWAIT so that Finish
is still responsible for reaping processes.

In DoWork, we call CheckDoneConsoleProcesses to do
the above, and only call ppoll/pselect on non-console
processes' fds. If there are only console processes,
then we block on an arbitrary process's completion;
if there are mixed processes, then we do not block.