Base completion on exit, not pipe closure
(POSIX only)

DoWork will now watch for SIGCHLD signals,
and mark subprocesses as done upon receipt.
Closure of a subprocess's pipe is no longer
taken to mean the subprocess is done. This
makes Ninja robust to tasks that close the
pipe early but continue to run, or fork and
hold onto the pipe but exit the immediate
subprocess of Ninja.

When a subprocess is done, we drain its
pipe by polling/reading until there is
nothing else buffered; this approximates
reading whatever the immediate subprocess
wrote before it exited.

Fixes #965
Obsoletes #966