Detach from terminal without creating a session.
Child processes must be detach from the terminal because they
run in background and cannot interact with the terminal.
setsid() actually does detach the process from the terminal but
also create a new session. Process in a new session cannot receive
SIGTSTP.

This patch fix an issue introduced by #909 that prevent #744 to
forward SIGTSTP to sub-processes when its received.