Fix NINJA_STATUS %r on dumb terminals
PR #999 made dumb terminals only output when edges finish.  PrintStatus
is called after finished_edges_ is incremented, which means the
calculation for running edges will always return 1 less than the real
number of running processes.  This happens on smart terminals too, but
ninja will immediately print the status for the next edge with
starting_edges_ incremented, so the incorrect value is never visible.

Pass a boolean specifying whether the status is being printed on an edge
finishing, and if so count the edge that just finished as being running.