Fix extra CR in subprocess output on Windows
Replace `\r\n` with `\n` before printing output from a subprocess.
This supports universal newline conversion in concert with the text mode
output of the C standard library that is performed by the `fwrite()`
call that is done in `PrintOrBuffer()`.  When `fwrite()` is called on
Windows, it will add a `\r` before any `\n` in the output being written.

Fixes #773