Add edge's most recent input mtime to ninja log for correct dirty status computation
If an edge's output files' mtimes are compared to the most recent
input's mtime, edges might be calculated as clean even if they are
actually dirty. While an edge is running its rule to produce its
outputs and an input to the edge is updated before the outputs are
written to disk, then subsequent runs will think that the outputs are
newer than the inputs, even though the inputs have actually been updated
and may be different than what were used to produce those outputs.

Instead of comparing output mtimes, ninja will add the most recent
input's mtime to the build log file when an edge completes if all
implicit dependencies were stat'd before the output began to build (all
implicit dependencies were already loaded via the deps log file). It
looks at both the explicit and implicit dependencies to get the most
recent input. On subsequent runs, ninja will use this value to compare
to the edge's most recent input's mtime to determine whether the outputs
are dirty if a valid mtime was recorded in the log file.