Add a new deps format for a newline-separate list of files.
Simplifies integration with scripts which no longer need to write a
Makefile format. The Makefile format is confusing in the presence of
multiple outputs, and has complex escaping and quoting rules.

The new list format has no escaping rules as the format is a newline-
separated list of literal filenames.