Use DETACHED_PROCESS to mitigate Windows close-process bug
Windows 10 has a close-process performance bug which can significantly
affect the performance of ninja builds. This bug causes process
destruction to be quite expensive (5-8 ms). Because process destruction
happens while holding a system-global lock it is serialized and it
causes micro-hangs in the UI (mouse pointer freezes for seconds at a
time, especially when many processes are terminating simultaneously).

This change specifies DETACHED_PROCESS when possible because this avoids
the expensive shutdown path, for some reason. It is effectively only
applied to compiler processes, so the problem is not entirely avoided,
but in some cases this fix is enough to improve build speeds by 20-30%.

Fixes #1283, at least partially