disk_interface.[cc,h]: Changed semantics for calling Stat()
Instead of trying to encode everything into a returned mtime, this
now returns false in the case of a failure (in which case the err
string will be set), and reports the mtime, and whether the file
exists or not, in the new StatResult struct parameter.

This allows all valid mtimes to be returned, including 0 and negative
timestamps.