Fix compilation under msys2 with native subprocess
The goal with this commit is to get the smart terminal to work with
msys2 bash, but keeping the rest as in the msvc version. This way,
subprocesses are spawned the same way, not using 'sh -c', in both
msys2 and msvc so the build.ninja files can be reused.

Signed-off-by: Fredrik Medley <fredrik.medley@gmail.com>