Silently ignore non-empty directories during 'clean'
Instead of returning an error if 'remove()' fails to remove a non-empty
directory let this pass as a success and silently ignore the directory. (A
Warning() can be added if deemed necessary).

This will allow targets that generates a lot of files, e.g. a docs directory
generated by Doxygen, to be used without having to list each file as a
byproduct or an explicit output for a rule.  Depending on the wrapper (sh,
.bat, or any other script-language) it can be varingly difficult to ignore the
error which will then propagate further up the build chain.

Using CMake as generator the following provides a solution for Doxygen though:

  add_custom_command(OUTPUT docs/Doxygen
    COMMAND ${CMAKE_COMMAND} -E make_directory docs
    COMMAND ${CMAKE_COMMAND} -E chdir docs doxygen -g)

  add_custom_command(OUTPUT docs/.generate
    COMMAND ${CMAKE_COMMAND} -E chdir docs doxygen
    COMMAND ${CMAKE_COMMAND} -E touch docs/.generate
    MAIN_DEPENDENCY docs/Doxygen)