Teach CanonicalizePath to use working dir for initial ../
With this change, given working dir '/a/b/proj', the path '../proj/foo'
will be translated into 'foo'. This sorts out the problem of paths in
deps files when using 'gcc -I ..' in projects with include files in a
structure like '#include "proj/foo/bar.h".

The problem is only visible for generated header files. If 'foo/bar.h'
is a source file, the underlying file system will map both 'foo/bar.h'
and '../proj/foo/bar.h' to the same file and mtime information for
ninja's build graph.

Using 'gcc -I <abs-path>' is not recommended, as described in the
manual, because '#include "proj/foo/bar.h"' in 'proj/xyz.c' generates
an absolute path in the dep file but '#include "foo/bar.h"' generates
a relative path. Absolute and relative paths will not be matched by
ninja and cause the same problem as described above, but only if
'foo/bar.h' is generated.