[wip/demo/do not merge] improve error messages in many cases
If last_token_ is a newline, the snippet would usually show an empty
line. In that case, show the previous line instead, which is the line
where the error would have been.

See the manifest_parser_test diff for what this changes.  Almost
everything is much better. Two things get slightly worse (the "expected
'command =' line for rules is now at the end of the rule line instead
of after it where it goes, and the same for the 'depth =' line for
pools). That's not ideal, but also not horrible.